feat: move google sign in logic to one file
