+++ src/screens/auth/SignUp.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react'
+import React, { useState } from 'react'
 import { View, TouchableWithoutFeedback } from 'react-native'
 import {
   Text,
@@ -14,7 +14,6 @@ import {
   AuthLayout,
   Avatar,
   FooterImages,
-  GoogleButton,
   //  TODO: temorary commented, as it is not implemented yet, remove comment when facebook access is ready
   // OtherAccess,
 } from '../../components/auth'
@@ -22,13 +21,11 @@ import { ShowSnackBar, dismissSnackBar } from '../../components/SnackBar'
 import emojis from '../../../emojis'
 import { useAuth } from '../../context/AuthContext'
 import { useTranslation } from '../../context/LanguageContext'
-import {
-  GoogleSignin,
-  statusCodes,
-} from '@react-native-community/google-signin'
+// TODO: Implement when "Sign-in in progress" problem has been fixed
+// import GoogleSignIn from './GoogleSignIn'
 
 const SignUp = ({ navigation }) => {
-  const { signUp, signInGoogle } = useAuth()
+  const { signUp } = useAuth()
   const styles = useStyleSheet(themedStyles)
   const [name, setName] = useState(null)
   const [email, setEmail] = useState(null)
@@ -37,36 +34,6 @@ const SignUp = ({ navigation }) => {
   const [loading, setLoading] = useState(false)
   const { user } = useTranslation()
   const showPasswordIcon = () => setSecureTextEntry(!secureTextEntry)
-
-  useEffect(() => {
-    GoogleSignin.configure({
-      webClientId:
-        '473036889467-niatcbo3v3bucf442gsfg2qle2dfhca5.apps.googleusercontent.com',
-      offlineAccess: true,
-      forceCodeForRefreshToken: true,
-    })
-  }, [])
-
-  const handleGoogleSignIn = async () => {
-    try {
-      setLoading(true)
-      await GoogleSignin.hasPlayServices()
-      const userInfo = await GoogleSignin.signIn()
-      await signInGoogle(userInfo.idToken)
-      setLoading(false)
-    } catch (error) {
-      setLoading(false)
-      if (error.code === statusCodes.SIGN_IN_CANCELLED) {
-        // user cancelled the login flow       
-      } else if (error.code === statusCodes.IN_PROGRESS) {
-        // operation (e.g. sign in) is in progress already
-      } else if (error.code === statusCodes.PLAY_SERVICES_NOT_AVAILABLE) {
-        // play services not available or outdated
-      } else {
-        console.error(error)
-      }
-    }
-  }
 
   const renderIcon = props => (
     <TouchableWithoutFeedback onPress={showPasswordIcon}>
@@ -146,9 +113,8 @@ const SignUp = ({ navigation }) => {
         <Button style={styles.button} onPress={handleSignUp}>
           {loading ? loadingSpinner : user.authentication.signUp}
         </Button>
-        <View style={styles.googleButton}>
-          <GoogleButton signIn={handleGoogleSignIn} disabled={loading} />
-        </View>
+        {/* TODO: Implement when "Sign-in in progress" problem has been fixed */}
+        {/* <GoogleSignIn /> */}
       </View>
       {/* TODO: temorary commented, as it is not implemented yet, remove comment when facebook access is ready */}
       {/* <OtherAccess /> */}
@@ -176,11 +142,6 @@ const themedStyles = StyleService.create({
     marginBottom: 5,
     backgroundColor: 'color-button-100',
     borderRadius: 5,
-    marginTop: 10,
-  },
-  googleButton: {
-    marginHorizontal: 50,
-    marginBottom: 5,
     marginTop: 10,
   },
   crendencials: {
